LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-only-unsafe-eval-expected.txt


PASS Eval is allowed because the CSP is report-only
FAIL SPV event is still raised assert_unreached: SPV event has not been received Reached unreachable code
FAIL Violation report status OK. assert_true: No CSP report sent, but expecting one. expected true got false
